Feature/CORE-271: warnings cleanup and config

Approved-by: Patrick Fromberg
Approved-by: Richard Mills

+++ src/test/bn_op_tests.cpp
@@ -369,7 +369,7 @@ BOOST_AUTO_TEST_CASE(bint_bint_numequalverify)
 BOOST_AUTO_TEST_CASE(operands_too_large)
 {
     GlobalConfig& config = GlobalConfig::GetConfig();
-    using test_args = tuple<int, int, opcodetype, bool, ScriptError>;
+    using test_args = tuple<uint64_t, uint64_t, opcodetype, bool, ScriptError>;
     const auto max_arg_len{ MAX_SCRIPT_NUM_LENGTH_AFTER_GENESIS };
 
     // set policy for script size, stack memory usage and max number length in scripts
